rectified message displayed when performing crud actions for departments to not include quotes in it.
rectified user not to be in more than one role.
rectified so that you cannot delete a user if there is a role attached to that user.
rectified so that you cannot delete a role if there are users or rights attached to that role.
rectified so that you cannot delete a right if there are roles attached to it.
add image in login background.
rectified message displayed when performing crud actions for departments to not include quotes in it.
rectified so that foreign key load only active records.